Moving back to the Apache license
