+++ util-lift/src/main/scala-2.10/com/outworkers/util/lift/JsonWrapper.scala
@@ -1,10 +1,23 @@
+/*
+ * Copyright 2013 - 2017 Outworkers Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.outworkers.util.lift
 
 import net.liftweb.json.JsonAST.JValue
 import net.liftweb.json.{pretty, render}
 
 object JsonWrapper {
-  def prettyRender(json: JValue): String = {
-    pretty(render(json))
-  }
+  def prettyRender(json: JValue): String = pretty(render(json))
 }
