+++ util-lift/src/test/scala/com/outworkers/util/lift/JsonExtractorsTest.scala
@@ -1,9 +1,23 @@
+/*
+ * Copyright 2013 - 2017 Outworkers Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.outworkers.util.lift
 
 import com.outworkers.util.testing._
 import net.liftweb.json.{DefaultFormats, JsonParser}
 import org.scalatest.{FlatSpec, Matchers}
-
 
 class JsonExtractorsTest extends FlatSpec with Matchers {
