--- util-lift/src/test/scala/com/outworkers/util/lift/Test.scala
@@ -1,3 +1,0 @@
-package com.outworkers.util.lift
-
-case class Test(name: String, amount: Int)
