+++ util-testing/src/test/scala/com/outworkers/util/testing/samples.scala
@@ -1,3 +1,18 @@
+/*
+ * Copyright 2013 - 2017 Outworkers Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.outworkers.util.testing
 
 import java.util.UUID
@@ -8,10 +23,6 @@ import java.util.UUID
   lastName: String,
   email: String
 )
-
-object User {
-  val x = 5
-}
 
 @sample case class CollectionSample(
   id: UUID,
