+++ util-validators/src/test/scala/com/outworkers/util/validators/ValidatorsTest.scala
@@ -1,4 +1,4 @@
-package com.outworkers.validators
+package com.outworkers.util.validators
 
 import cats.Apply
 import cats.data.Validated.{Invalid, Valid}
